Status command indicates missing scripts.
Add some text in the status description field to indicate if a script file is
not found in the file system. This can happen, for example, when running
status on a database when the local environment is on a branch that does not
include one of the migrations previously applied to the database.